+++ spec/spec_helper.rb
@@ -13,10 +13,3 @@ RSpec.configure do |config|
     config.mock_with :mocha
 end
 
-# Borrowed from YARD spec helper
-def parse_file(file, thisfile = __FILE__, log_level = log.level, ext = '.pp')
-  Registry.clear
-  path = File.join(File.dirname(thisfile), 'examples', file.to_s + ext)
-  YARD::Parser::SourceParser.parse(path, [], log_level)
-end
-
